# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API LiqaaSpace",
        "description": "Documentation de l'API LiqaaSpace",
        "contact": {
            "name": "Équipe LiqaaSpace",
            "email": "contact@liqaaspace.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/disponibilites": {
            "get": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Liste toutes les disponibilités",
                "operationId": "c955c9d7a1d13248ecd6274cb2db3a70",
                "responses": {
                    "200": {
                        "description": "Liste des disponibilités récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Créer une nouvelle disponibilité",
                "operationId": "57706ec08f0d0d4d9b76e3b3239b38cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "salle_id",
                                    "date",
                                    "heure_debut",
                                    "heure_fin",
                                    "statut"
                                ],
                                "properties": {
                                    "salle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-22"
                                    },
                                    "heure_debut": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "heure_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "occupé"
                                        ],
                                        "example": "disponible"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Disponibilité créée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/disponibilites/{id}": {
            "get": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Afficher une disponibilité spécifique",
                "operationId": "d7c727bdf060029ca1e4f93c6d42f2cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilité",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilité récupérée avec succès"
                    },
                    "404": {
                        "description": "Disponibilité non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Mettre à jour une disponibilité",
                "operationId": "0ea006c3eec361bf1eda27c4f246d8ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilité à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "salle_id",
                                    "date",
                                    "heure_debut",
                                    "heure_fin",
                                    "statut"
                                ],
                                "properties": {
                                    "salle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-22"
                                    },
                                    "heure_debut": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "heure_fin": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "disponible",
                                            "occupé"
                                        ],
                                        "example": "occupé"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disponibilité mise à jour avec succès"
                    },
                    "404": {
                        "description": "Disponibilité non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Disponibilités"
                ],
                "summary": "Supprimer une disponibilité",
                "operationId": "1363809bbd9388368b28ffe0b1f6d210",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilité à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilité supprimée avec succès"
                    },
                    "404": {
                        "description": "Disponibilité non trouvée"
                    }
                }
            }
        },
...
    },
...
}